Add integration test for send_est
