+++ erica/application/JobService/job.py
@@ -35,8 +35,8 @@ async def perform_job(request_id: UUID, repository: base_repository_interface, s
         try:
             response = await service.apply_to_elster(request_payload, True)
             # We do not want to send the server_response or eric_response to the clients in the success case
-            response['server_response'] = None
-            response['eric_response'] = None
+            response.pop('server_response', None)
+            response.pop('eric_response', None)
             entity.result = response
             entity.status = Status.success
             repository.update(entity.id, entity)
